requirements.txt
aiohttp==3.9.0b0
APScheduler
async-timeout
chardet
discord.py
emoji
multidict
pytz
six
tzlocal
websockets
numpy
matplotlib
python-dotenv
